pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "degen"
version = "0.5.3"
description = "Data Engineering Architecture Composer CLI"
authors = [{ name="Felipe Castro" }]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "typer>=0.9.0",
    "questionary>=2.0.1",
    "rich>=13.7.0",
    "PyYAML>=6.0"
]

[project.scripts]
degen = "degen.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["degen*"]